unilabs_csv_to_genesys-main/requirements.txt
annotated-types==0.7.0
pydantic==2.7.1
pydantic_core==2.18.2
typing_extensions==4.12.0
pytest==8.2.1
PyYAML~=6.0.1
SQLAlchemy~=2.0.30
pyodbc~=5.1.0
aiosmtplib~=3.0.1